추가: 댓글삭제ajax 변경중
.

+++ src/main/java/com/breaktheice/moimat/domain/BasicDomain.java
@@ -8,21 +8,21 @@ import lombok.Data;
 public class BasicDomain {
 	
 	//team post테이블
-	private int postId;
-	private int teamId;
-	private int brdId;
+	private long postId;
+	private long teamId;
+	private long brdId;
 	private String postTitle;
 	private String postContent;
-	private int tmemId;
-	private int postTmemLevel;
+	private long tmemId;
+	private long postTmemLevel;
 	private String postNickname;
 	private String postEmail;
 	private Timestamp postRegdate;
 	private Timestamp postUpdate;
-	private int postHit;
+	private long postHit;
 	
 	//team_comments 댓글 테이블
-	private int cmtId;
+	private long cmtId;
 	private String cmtContent;
 	private String cmtNickname;
 	private String cmtEmail;
